+++ app/lib/l10n/app_localizations_de.dart
@@ -4208,6 +4208,42 @@ class AppLocalizationsDe extends AppLocalizations {
   String get danceEditorMonthDec => 'Dez.';
 
   @override
+  String get monthFullJanuary => 'January';
+
+  @override
+  String get monthFullFebruary => 'February';
+
+  @override
+  String get monthFullMarch => 'March';
+
+  @override
+  String get monthFullApril => 'April';
+
+  @override
+  String get monthFullMay => 'May';
+
+  @override
+  String get monthFullJune => 'June';
+
+  @override
+  String get monthFullJuly => 'July';
+
+  @override
+  String get monthFullAugust => 'August';
+
+  @override
+  String get monthFullSeptember => 'September';
+
+  @override
+  String get monthFullOctober => 'October';
+
+  @override
+  String get monthFullNovember => 'November';
+
+  @override
+  String get monthFullDecember => 'December';
+
+  @override
   String get danceEditorAddTuneHint => 'Vorgeschlagene Melodie hinzufügen…';
 
   @override