+++ app/lib/l10n/app_localizations_nl.dart
@@ -4186,6 +4186,42 @@ class AppLocalizationsNl extends AppLocalizations {
   String get danceEditorMonthDec => 'Dec';
 
   @override
+  String get monthFullJanuary => 'January';
+
+  @override
+  String get monthFullFebruary => 'February';
+
+  @override
+  String get monthFullMarch => 'March';
+
+  @override
+  String get monthFullApril => 'April';
+
+  @override
+  String get monthFullMay => 'May';
+
+  @override
+  String get monthFullJune => 'June';
+
+  @override
+  String get monthFullJuly => 'July';
+
+  @override
+  String get monthFullAugust => 'August';
+
+  @override
+  String get monthFullSeptember => 'September';
+
+  @override
+  String get monthFullOctober => 'October';
+
+  @override
+  String get monthFullNovember => 'November';
+
+  @override
+  String get monthFullDecember => 'December';
+
+  @override
   String get danceEditorAddTuneHint => 'Een voorgesteld deuntje toevoegen…';
 
   @override